Adds support for react-instantsearch-nextjs
Integrates the react-instantsearch-nextjs library to enhance search functionality within the application.

Updates the import structure and modifies components to utilize the new InstantSearchNext component.

Enhances the mobile filter component to track selected categories and display the count of selected filters.

Sets the page to dynamic loading behavior for improved performance.